+++ packages/elements/select/examples/01-multi-select.js
@@ -4,6 +4,8 @@ import { Select } from './common/components';
 import { cities } from './common/data';
 
 // data imported for brevity; equal to the options from Single Select example
-export default () => (
+const MultiExample = () => (
   <Select options={cities} isMulti placeholder="Choose a City" />
 );
+
+export default MultiExample;