+++ packages/elements/select/examples/common/components.js
@@ -2,6 +2,7 @@
 import { withValue } from 'react-value';
 import BaseSelect, {
   CheckboxSelect as BaseCheckboxSelect,
+  CountrySelect as BaseCountrySelect,
   RadioSelect as BaseRadioSelect,
 } from '../../src';
 
@@ -11,6 +12,9 @@ Select.displayName = 'ExampleSelectWithValue';
 
 export const CheckboxSelect = withValue(BaseCheckboxSelect);
 CheckboxSelect.displayName = 'ExampleCheckboxSelectWithValue';
+
+export const CountrySelect = withValue(BaseCountrySelect);
+CountrySelect.displayName = 'ExampleCountrySelectWithValue';
 
 export const RadioSelect = withValue(BaseRadioSelect);
 RadioSelect.displayName = 'ExampleRadioSelectWithValue';