+++ packages/fabric/editor-core/__tests__/plugins/media/keymap.ts
@@ -84,9 +84,9 @@ describe('media - keymaps', () => {
     });
   });
 
-  // TODO: @see ED-3682
-  describe.skip('Shift-Enter keypress', () => {
-    it('splits media group', () => {
+  describe('Shift-Enter keypress', () => {
+    // Wait for ED-3741 to refactor
+    it.skip('splits media group', () => {
       const { editorView, pluginState } = editor(doc(p('{<>}')));
       const splitMediaGroupSpy = jest.spyOn(pluginState, 'splitMediaGroup');