+++ packages/elements/banner/CHANGELOG.md
@@ -1,5 +1,8 @@
 # @atlaskit/banner
 
+## 4.1.3
+- [patch] Packages Flow types for elements components [3111e74](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/3111e74)
+
 ## 4.1.2
 - [patch] Flatten examples for easier consumer use [145b632](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/145b632)