+++ packages/elements/empty-state/CHANGELOG.md
@@ -1,5 +1,8 @@
 # @atlaskit/empty-state
 
+## 0.2.1
+- [patch] Packages Flow types for elements components [3111e74](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/3111e74)
+
 ## 0.2.0
 - [minor] Add React 16 support. [12ea6e4](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/12ea6e4)